class07-20250601/04_agents_as_tool/chatbot_agent_as_tool/pyproject.toml
[project]
name = "chatbot-agent-as-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
